Update old locales
Very old versions of Java (before Java 1.4) used in_ID for the Indonesian language, following an older ISO standard.
If you create a Locale with "in", "ID", Java will automatically convert it to "id_ID".

iw_IL and he_IL refer to the same locale, but with a historical difference:
iw was the language code used for Hebrew in older versions of Java (before Java 7).
he is the updated and standard ISO 639-1 code for Hebrew.